machmiking

+++ app/src/main/java/com/example/snapking/modelo/Sala.kt
@@ -1,7 +1,8 @@
 package com.example.snapking.modelo
 
-data class Sala(var nombre:String, var capacidad:Int,
-                var anfitrion:String, var estado:Boolean, var ronda: Ronda?, var rondas_totales:Int, var jugadores:List<Jugador>){
+data class Sala(
+    var nombre:String, var capacidad: Long,
+    var anfitrion:String, var estado:Boolean, var ronda: Ronda?, var rondas_totales: Long, var jugadores:List<Jugador>){
 
 
 }
